ai_service/requirements.txt
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.4.2
fastapi==0.104.1
uvicorn==0.24.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0 